fix: building troubleshooting.js from typescript, moving source files to src/js/
